refactor to route and sessionservice
