+++ app/routes/agenda/agendaitems/agendaitem.js
@@ -1,5 +1,7 @@
 import Route from '@ember/routing/route';
 import { inject } from '@ember/service';
+import moment from 'moment';
+
 
 export default Route.extend({
   sessionService: inject(),
@@ -11,11 +13,52 @@ export default Route.extend({
     });
   },
 
-  afterModel(model) {
+  async afterModel(model) {
+    const subcase = await model.get('subcase');
+    const result = await this.shouldShowEditedWarning(subcase);
     this.set('sessionService.selectedAgendaItem', model);
+    this.set('sessionService.showNewsItemIsEditedWarning', result);
   },
 
-  actions: {
+  async shouldShowEditedWarning(subcase) {
+    const newsletterInfoForSubcase = await subcase.get('newsletterInfo');
+    const documents = await subcase.get('documents');
+    if (!documents) {
+      return;
+    }
+    const documentsOfTypeNota = [];
+    await Promise.all(documents.map(document => {
+      return document.get('type').then((type) => {
+        if (type.label === 'Nota') {
+          documentsOfTypeNota.push(document);
+        }
+      })
+    }));
+
+    const lastVersionOfNotas = await Promise.all(documentsOfTypeNota.map((nota) => {
+      return nota.get('lastDocumentVersion');
+    }));
+
+    const mostRecentlyAddedNotaDocumentVersion = lastVersionOfNotas.sortBy('lastModified').lastObject;
+    const modifiedDateFromMostRecentlyAddedNotaDocumentVersion = mostRecentlyAddedNotaDocumentVersion.modified;
+
+    const newsletterInfoOnSubcaseLastModifiedTime = newsletterInfoForSubcase.modified;
+    if (newsletterInfoOnSubcaseLastModifiedTime) {
+      if (moment(newsletterInfoOnSubcaseLastModifiedTime).isBefore(moment(modifiedDateFromMostRecentlyAddedNotaDocumentVersion))) {
+        return true;
+        // this.set('showNewsItemIsEditedWarning', true);
+      } else {
+        return false;
+        // this.set('showNewsItemIsEditedWarning', false);
+      }
+    } else {
+      return true;
+      // this.set('showNewsItemIsEditedWarning', true);
+    }
+  },
+
+
+    actions: {
     refreshRoute() {
       this._super(...arguments);
       this.refresh();
